Add files via upload
This is a basic table dev in html and styled with an external css.
In this work, I write a simple html code to create a 6 row table inclusive of a header, with 3 columns.
In the same work, I have another table created just getting my hands on rowspan and colspan. Quite interesting.
Adding a little catchy something, I introduce the nth-child pseudo-class to my table in the css code.